+++ lib/extend/signature.ts
@@ -25,7 +25,7 @@ function signatureExtend(): () => void {
 
         // 校验该扩展是否已添加过
         if (this._signatureExtendAdded) {
-            console.error('`signatureExtend` can only be added to ajax once!');
+            console && console.error('`signatureExtend` can only be added to ajax once!');
         }
 
         // 添加标志符用来校验该扩展是否已添加