+++ lib/index.ts
@@ -2,4 +2,6 @@ import ajax from './ajax';
 
 export * as Ajax from './interface';
 
+export { default as AjaxBase } from './ajax/base';
+
 export default ajax;