+++ types/index.d.ts
@@ -1,3 +1,4 @@
 import ajax from './ajax';
 export * as Ajax from './interface';
+export { default as AjaxBase } from './ajax/base';
 export default ajax;