+++ types/utils/catch.d.ts
@@ -1,0 +1,9 @@
+import * as IAjax from '../interface';
+export declare function catchAjaxError({ e, method, url, params, callback, type, }: {
+    e: Error;
+    method: IAjax.IMethod;
+    url: string;
+    params?: IAjax.IParams;
+    callback: IAjax.ICatchError;
+    type?: 'uncaught' | 'log';
+}): void;