+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 <font color=red><strong>删除</strong></font>
 
 
+# 1.2.0
+> 2020.05.25  
+1. <font color=blue>增强</font> `config`配置项添加`catchError`用来捕获错误，记录日志
+2. <font color=blue>增强</font> 暴露`AjaxBase`
+
+
 # 1.1.0
 > 2020.05.25  
 1. <font color=blue>增强</font> `config`配置项添加支持`beforeSend`和`processData`